[titusan] adding functionality for when a user clicks the friends icon at the top right of the watch screen
