+++ FullCrossApplication(iOS)/Views/TabViews/WatchScreen.swift
@@ -83,6 +83,8 @@ struct WatchScreen: View {
         }
         .sheet(isPresented: $showSocialDialog) {
             SocialConnectionDialog(
+                contactsViewModel: contactsViewModel,
+                notificationsViewModel: notificationsViewModel,
                 onDismiss: { showSocialDialog = false }
             )
         }
@@ -309,302 +311,5 @@ struct UpcomingStreamCard: View {
         } catch {
             print("Error saving event: \(error.localizedDescription)")
         }
-    }
-}
-
-struct SocialConnectionDialog: View {
-    let onDismiss: () -> Void
-    
-    @StateObject private var contactsViewModel: ContactsViewModel
-    @EnvironmentObject private var notificationsViewModel: NotificationsViewModel
-    @State private var searchQuery = ""
-    @State private var isSyncExpanded = false
-    @State private var isFindFriendsExpanded = false
-    @Environment(\.openURL) private var openURL
-    
-    init(onDismiss: @escaping () -> Void) {
-        self.onDismiss = onDismiss
-        // Initialize ContactsViewModel with AuthViewModel
-        _contactsViewModel = StateObject(wrappedValue: ContactsViewModel(authViewModel: AuthViewModel()))
-    }
-    
-    var body: some View {
-        NavigationView {
-            ScrollView {
-                VStack(spacing: 16) {
-                    // Friend Requests Section
-                    friendRequestsSection
-                    
-                    // Find Friends Section
-                    findFriendsSection
-                    
-                    Divider()
-                    
-                    // Sync Contacts Section
-                    syncContactsSection
-                    
-                    // Refer Friends Button
-                    referFriendsButton
-                }
-                .padding()
-            }
-            .navigationTitle("Connect with Friends")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .navigationBarTrailing) {
-                    Button("Done") {
-                        onDismiss()
-                    }
-                }
-            }
-        }
-    }
-    
-    private var friendRequestsSection: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text("Friend Requests")
-                .font(.headline)
-            
-            ForEach(notificationsViewModel.notifications.filter { $0.type == .friendRequest && !$0.read }) { request in
-                FriendRequestCard(
-                    request: request,
-                    contactsViewModel: contactsViewModel,
-                    notificationsViewModel: notificationsViewModel
-                )
-            }
-            
-            Divider()
-                .padding(.vertical, 8)
-        }
-    }
-    
-    private var findFriendsSection: some View {
-        VStack(spacing: 8) {
-            DisclosureGroup(
-                isExpanded: $isFindFriendsExpanded,
-                content: {
-                    VStack(spacing: 12) {
-                        SearchBar(text: $searchQuery)
-                            .onChange(of: searchQuery) { newValue in
-                                Task {
-                                    await contactsViewModel.searchUsers(newValue)
-                                }
-                            }
-                        
-                        if contactsViewModel.isSearching {
-                            ProgressView()
-                                .padding()
-                        } else if !contactsViewModel.searchResults.isEmpty {
-                            ForEach(contactsViewModel.searchResults) { user in
-                                SearchResultRow(
-                                    user: user,
-                                    contactsViewModel: contactsViewModel
-                                )
-                            }
-                        } else if searchQuery.count >= 2 {
-                            Text("No users found")
-                                .foregroundColor(.secondary)
-                                .padding()
-                        }
-                    }
-                },
-                label: {
-                    Label("Find Friends", systemImage: "magnifyingglass")
-                        .font(.headline)
-                }
-            )
-        }
-    }
-    
-    private var syncContactsSection: some View {
-        VStack(spacing: 8) {
-            DisclosureGroup(
-                isExpanded: $isSyncExpanded,
-                content: {
-                    if contactsViewModel.isLoading {
-                        ProgressView()
-                            .padding()
-                    } else if !contactsViewModel.contacts.isEmpty {
-                        ForEach(contactsViewModel.contacts) { contact in
-                            ContactRow(contact: contact)
-                        }
-                    }
-                },
-                label: {
-                    Label("Sync Contacts", systemImage: "person.crop.circle.badge.plus")
-                        .font(.headline)
-                }
-            )
-            .onChange(of: isSyncExpanded) { expanded in
-                if expanded && contactsViewModel.contacts.isEmpty {
-                    Task {
-                        await contactsViewModel.syncContacts()
-                    }
-                }
-            }
-        }
-    }
-    
-    private var referFriendsButton: some View {
-        Button {
-            let message = "Join me on our church app! Download it here: [Your App Link]"
-            let url = URL(string: "sms:&body=\(message.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? "")")
-            if let url = url {
-                openURL(url)
-            }
-        } label: {
-            Label("Refer Friends", systemImage: "square.and.arrow.up")
-                .frame(maxWidth: .infinity)
-                .padding()
-                .background(Color.accentColor)
-                .foregroundColor(.white)
-                .clipShape(RoundedRectangle(cornerRadius: 8))
-        }
-    }
-}
-
-struct SearchBar: View {
-    @Binding var text: String
-    
-    var body: some View {
-        HStack {
-            Image(systemName: "magnifyingglass")
-                .foregroundColor(.secondary)
-            
-            TextField("Search by name or phone", text: $text)
-                .textFieldStyle(RoundedBorderTextFieldStyle())
-            
-            if !text.isEmpty {
-                Button(action: { text = "" }) {
-                    Image(systemName: "xmark.circle.fill")
-                        .foregroundColor(.secondary)
-                }
-            }
-        }
-        .padding(.horizontal)
-    }
-}
-
-struct FriendRequestCard: View {
-    let request: Notification
-    let contactsViewModel: ContactsViewModel
-    let notificationsViewModel: NotificationsViewModel
-    
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                Text(request.fromUserName)
-                    .font(.headline)
-                Text("Sent you a friend request")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-            }
-            
-            Spacer()
-            
-            HStack {
-                Button("Accept") {
-                    Task {
-                        await contactsViewModel.acceptFriendRequest(request.fromUserId)
-                        await notificationsViewModel.markAsRead(request.id)
-                    }
-                }
-                .buttonStyle(.borderedProminent)
-                
-                Button("Decline") {
-                    Task {
-                        await contactsViewModel.declineFriendRequest(request.fromUserId)
-                        await notificationsViewModel.markAsRead(request.id)
-                    }
-                }
-                .buttonStyle(.bordered)
-            }
-        }
-        .padding()
-        .background(Color(.systemGray6))
-        .clipShape(RoundedRectangle(cornerRadius: 8))
-    }
-}
-
-struct SearchResultRow: View {
-    let user: UserProfile
-    let contactsViewModel: ContactsViewModel
-    
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                Text(user.fullName)
-                    .font(.headline)
-                if !user.phoneNumber.isEmpty {
-                    Text(user.phoneNumber)
-                        .font(.subheadline)
-                        .foregroundColor(.secondary)
-                }
-            }
-            
-            Spacer()
-            
-            switch user.friendshipStatus {
-            case .none:
-                Button("Add Friend") {
-                    Task {
-                        await contactsViewModel.sendFriendRequest(toUserId: user.id, toUserName: user.fullName)
-                    }
-                }
-                .foregroundColor(.green)
-                
-            case .pending:
-                Text("Pending")
-                    .foregroundColor(.secondary)
-                
-            case .accepted:
-                Image(systemName: "checkmark.circle.fill")
-                    .foregroundColor(.green)
-                
-            case .declined:
-                Button("Try Again") {
-                    Task {
-                        await contactsViewModel.sendFriendRequest(toUserId: user.id, toUserName: user.fullName)
-                    }
-                }
-            }
-        }
-        .padding(.horizontal)
-    }
-}
-
-struct ContactRow: View {
-    let contact: Contact
-    @Environment(\.openURL) private var openURL
-    
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                Text(contact.name)
-                    .font(.headline)
-                if let phone = contact.phoneNumber {
-                    Text(phone)
-                        .font(.subheadline)
-                        .foregroundColor(.secondary)
-                }
-            }
-            
-            Spacer()
-            
-            if contact.isAppUser {
-                Image(systemName: "checkmark.circle.fill")
-                    .foregroundColor(.green)
-            } else {
-                Button("Invite") {
-                    let message = "Join me on our church app! Download it here: [Your App Link]"
-                    if let phone = contact.phoneNumber,
-                       let url = URL(string: "sms:\(phone)&body=\(message.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? "")") {
-                        openURL(url)
-                    }
-                }
-                .buttonStyle(.bordered)
-            }
-        }
-        .padding(.horizontal)
     }
 }
